+++ packages/shadow-objects/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to [@spearwolf/shadow-objects](https://github.com/spearwolf/
 The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.26.3] - 2026-01-15
+
+- fix type definitions for `provideContext()`, `provideGlobalContext()`, `useContext()`, `useParentContext()` and `useProperty()` when using the deprecated third argument as `isEqual` callback
+- improve type definitions for `createResource()`
+  - resource is set to `undefined` after entity destruction
+
 ## [0.26.2] - 2026-01-12
 
 - fix the check for when the deprecation warning is shown in `provideContext()` and `provideGlobalContext()`